Reorganisation of module to better fit to the TS lib
