+++ dashipy/dashipy/components/box.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from dashipy.lib import Container
+from dashipy import Container
 
 
 @dataclass(frozen=True)
