+++ dashipy/dashipy/controllers/contributions.py
@@ -1,5 +1,5 @@
-from .extensioncontext import ExtensionContext
-from .response import Response
+from dashipy.extensioncontext import ExtensionContext
+from dashipy.response import Response
 
 
 def get_contributions(ext_ctx: ExtensionContext) -> Response:
