--- dashipy/dashipy/lib/__init__.py
@@ -1,7 +1,0 @@
-from .callback import Callback
-from .callback import Input
-from .callback import Output
-from .component import Component
-from .container import Container
-from .extension import Contribution
-from .extension import Extension
