+++ dashipy/my_extension/__init__.py
@@ -1,4 +1,4 @@
-from dashipy.lib.extension import Extension
+from dashipy import Extension
 from .my_panel_1 import panel as my_panel_1
 from .my_panel_2 import panel as my_panel_2
