+++ dashipy/my_extension/my_panel_1.py
@@ -1,10 +1,10 @@
 import plotly.graph_objects as go
 from plotly.graph_objs import Layout
 
-from dashipy.components import Plot, Box, Dropdown
+from dashipy import (Component, Input, Output)
+from dashipy.components import (Plot, Box, Dropdown)
 from dashipy.demo.contribs import Panel
 from dashipy.demo.context import Context
-from dashipy.lib import Output, Input, Component
 
 
 panel = Panel(__name__, title="Panel A")
