+++ dashipy/tests/contribs/panel_test.py
@@ -1,7 +1,7 @@
 import unittest
 
 from dashipy.demo.contribs import Panel
-from dashipy.lib.callback import Input, Callback, Output
+from dashipy.callback import Input, Callback, Output
 
 
 def my_callback_1(ctx):
